+++ src/RemoteRoute/Downloader/DownloaderClient.cpp
@@ -33,7 +33,7 @@ bool RDTcpClient::IsStopped() const
 	return m_bStopped;
 }
 
-__int64 RDTcpClient::GetSecondsSinceLastRcv() const
+long long RDTcpClient::GetSecondsSinceLastRcv() const
 {
 	auto cur = std::chrono::duration_cast<std::chrono::seconds>
 		(std::chrono::system_clock::now().time_since_epoch()).count();
